+++ src/components/Banner.tsx
@@ -2,7 +2,12 @@ import CountUp from "react-countup"
 import { FaPlus, FaRegFolderOpen } from "react-icons/fa6"
 import { Link } from "react-router"
 
-const Banner = () => {
+type BannerProps = {
+    total: number
+    path: string
+}
+
+const Banner = ({total, path}: BannerProps) => {
     return (
         <div className="rounded-lg bg-slate-50 border border-blue-100 text-blue-500 py-6 px-10">
             <div className="flex justify-between items-center">
@@ -13,11 +18,11 @@ const Banner = () => {
                         <div className="p-2 rounded-full bg-white border border-blue-300">
                             <FaRegFolderOpen className="text-blue-500 text-2xl" />
                         </div>
-                        <CountUp start={0} end={500} duration={2} className="text-3xl font-bold" />
+                        <CountUp start={0} end={total} duration={2} className="text-3xl font-bold" />
                     </div>
                 </div>
                 <div className="flex gap-3 items-center">
-                    <Link to="/dashbord/projects/create" className="py-2 px-5 bg-blue-500 rounded-lg text-blue-50 flex items-center gap-2">
+                    <Link to={path} className="py-2 px-5 bg-blue-500 rounded-lg text-blue-50 flex items-center gap-2">
                         <FaPlus size={14} />
                         <span>Add</span>
                     </Link>