+++ src/layouts/home/HomeLayout.tsx
@@ -1,0 +1,9 @@
+const HomeLayout = () => {
+    return (
+        <div className="w-full bg-slate-50">
+            
+        </div>
+    )
+}
+
+export default HomeLayout