+++ src/pages/admin/projects/ProjectListPage.tsx
@@ -5,7 +5,7 @@ const ProjectListPage = () => {
         <div className="">
             <h1 className="">Keep track your project so you can make a portfolio here</h1>
             <div className="p-2 mt-12">
-                <Banner />
+                <Banner total={5} path="/dashboard/projects/create" />
             </div>
         </div>
     )