+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 - Add Sidebar Component
 - Add Authentication Page
 - Add Project Page
+- Add login and register hooks